backend/requirements.txt
asgiref==3.9.1
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
fonttools==4.58.0
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
matplotlib==3.10.3
msgpack==1.1.1
myModule==1.0.0
numba==0.61.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pycparser==2.22
pygame==2.5.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.4.3
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2022.7
urllib3==2.5.0
